# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PoolTrace API",
    "description": "Local mortgage analytics and reconciliation. Demo data is synthetic; commercial adapters require authorized input.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/pools/{pool_id}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_api_v1_pools__pool_id__history_get",
        "parameters": [
          {
            "name": "pool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pool Id"
            }
          },
          {
            "name": "knowledge_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pools/{pool_id}/versions": {
      "get": {
        "summary": "Versions",
        "operationId": "versions_api_v1_pools__pool_id__versions_get",
        "parameters": [
          {
            "name": "pool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pools/{pool_id}/analyze": {
      "post": {
        "summary": "Run Analysis",
        "operationId": "run_analysis_api_v1_pools__pool_id__analyze_post",
        "parameters": [
          {
            "name": "pool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pool Id"
            }
          },
          {
            "name": "knowledge_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge At"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pools/{pool_id}/reconcile": {
      "post": {
        "summary": "Run Reconciliation",
        "operationId": "run_reconciliation_api_v1_pools__pool_id__reconcile_post",
        "parameters": [
          {
            "name": "pool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pool Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pools/{pool_id}/prepayment": {
      "get": {
        "summary": "Prepayment",
        "operationId": "prepayment_api_v1_pools__pool_id__prepayment_get",
        "parameters": [
          {
            "name": "pool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pool Id"
            }
          },
          {
            "name": "knowledge_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest": {
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_api_v1_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/references": {
      "post": {
        "summary": "Reference",
        "operationId": "reference_api_v1_references_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReferenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/freddie/performance": {
      "post": {
        "summary": "Import Freddie",
        "operationId": "import_freddie_api_v1_freddie_performance_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerformanceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Read Freddie",
        "operationId": "read_freddie_api_v1_freddie_performance_get",
        "parameters": [
          {
            "name": "knowledge_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge At"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/breaks/{break_id}": {
      "patch": {
        "summary": "Update Break",
        "operationId": "update_break_api_v1_breaks__break_id__patch",
        "parameters": [
          {
            "name": "break_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Break Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tolerances": {
      "put": {
        "summary": "Set Tolerances",
        "operationId": "set_tolerances_api_v1_tolerances_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToleranceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events": {
      "get": {
        "summary": "Events",
        "operationId": "events_api_v1_events_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/export/{pool_id}": {
      "get": {
        "summary": "Export",
        "operationId": "export_api_v1_export__pool_id__get",
        "parameters": [
          {
            "name": "pool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pool Id"
            }
          },
          {
            "name": "knowledge_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Knowledge At"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/demo/{action}": {
      "post": {
        "summary": "Demo Action",
        "operationId": "demo_action_api_v1_demo__action__post",
        "parameters": [
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BreakUpdate": {
        "properties": {
          "owner": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Owner"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "investigating",
              "resolved"
            ],
            "title": "Status"
          },
          "note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "owner",
          "status",
          "note"
        ],
        "title": "BreakUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestRequest": {
        "properties": {
          "source": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Source",
            "default": "canonical-upload"
          },
          "csv_text": {
            "type": "string",
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Csv Text"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "csv_text"
        ],
        "title": "IngestRequest"
      },
      "PerformanceRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 2000000,
            "minLength": 1,
            "title": "Text"
          },
          "release": {
            "type": "string",
            "enum": [
              "r47",
              "pre-r47"
            ],
            "title": "Release",
            "default": "r47"
          },
          "source": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Source",
            "default": "freddie-sfll"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "PerformanceRequest"
      },
      "ReferenceRequest": {
        "properties": {
          "pool_id": {
            "type": "string",
            "title": "Pool Id"
          },
          "source": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source"
          },
          "as_of": {
            "type": "string",
            "format": "date",
            "title": "As Of"
          },
          "price": {
            "type": "number",
            "maximum": 300.0,
            "exclusiveMinimum": 0.0,
            "title": "Price"
          },
          "wal": {
            "type": "number",
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Wal"
          },
          "duration": {
            "type": "number",
            "maximum": 100.0,
            "minimum": -100.0,
            "title": "Duration"
          },
          "oas_bps": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 5000.0,
                "minimum": -5000.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Oas Bps"
          },
          "cpr": {
            "type": "number",
            "maximum": 0.8,
            "minimum": 0.0,
            "title": "Cpr"
          },
          "discount_rate": {
            "type": "number",
            "maximum": 0.25,
            "minimum": 0.0,
            "title": "Discount Rate"
          },
          "conventions": {
            "type": "string",
            "maxLength": 300,
            "title": "Conventions",
            "default": "monthly-net-coupon;current-balance;flat-continuous;no-accrual"
          },
          "duration_type": {
            "type": "string",
            "maxLength": 80,
            "title": "Duration Type",
            "default": "fixed_cashflow_parallel_rate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pool_id",
          "source",
          "as_of",
          "price",
          "wal",
          "duration",
          "cpr",
          "discount_rate"
        ],
        "title": "ReferenceRequest"
      },
      "Scenario": {
        "properties": {
          "cpr": {
            "type": "number",
            "maximum": 0.8,
            "minimum": 0.0,
            "title": "Cpr",
            "default": 0.08
          },
          "discount_rate": {
            "type": "number",
            "maximum": 0.25,
            "minimum": 0.0,
            "title": "Discount Rate",
            "default": 0.045
          },
          "market_price": {
            "type": "number",
            "maximum": 200.0,
            "exclusiveMinimum": 20.0,
            "title": "Market Price",
            "default": 100.0
          },
          "volatility": {
            "type": "number",
            "maximum": 0.08,
            "minimum": 0.0,
            "title": "Volatility",
            "default": 0.01
          },
          "paths": {
            "type": "integer",
            "maximum": 512.0,
            "minimum": 32.0,
            "title": "Paths",
            "default": 128
          },
          "seed": {
            "type": "integer",
            "maximum": 4294967295.0,
            "minimum": 0.0,
            "title": "Seed",
            "default": 17
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Scenario"
      },
      "ToleranceUpdate": {
        "properties": {
          "price": {
            "type": "number",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Price"
          },
          "wal": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Wal"
          },
          "duration": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Duration"
          },
          "oas_bps": {
            "type": "number",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Oas Bps"
          }
        },
        "type": "object",
        "required": [
          "price",
          "wal",
          "duration",
          "oas_bps"
        ],
        "title": "ToleranceUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}